feat: 更新过滤内容

+++ model/blamer/dep_blamer.py
@@ -307,21 +307,24 @@ def get_entity_commits(repo_path: str, accompany_dep: str, old_base_commits: str
     filter_file_set = set(filter(is_accompany_file, file_set))
     blame = create_blamer(blame_dict)
     print(' get entities commits')
-    for ent in ents:
-        if ent.path not in filter_file_set:
-            continue
-        ent_commits = blame(ent.path, ent.start_line, ent.end_line)
-        ownership_data.append(EntOwnership(ent, ent_commits))
-    with open(f"{out_path}/ownership.csv", "w", encoding="utf-8", newline="") as f:
-        dump_ownership(ownership_data, f)
+    if os.path.exists(f"{out_path}/ownership.csv"):
+        print(' entities commits existed')
+    else:
+        for ent in ents:
+            if ent.path not in filter_file_set:
+                continue
+            ent_commits = blame(ent.path, ent.start_line, ent.end_line)
+            ownership_data.append(EntOwnership(ent, ent_commits))
+        with open(f"{out_path}/ownership.csv", "w", encoding="utf-8", newline="") as f:
+            dump_ownership(ownership_data, f)
 
-    # with open(f"{analyzed_root}_ownership_lineageos_{sha}.csv", "w", newline="") as f:
-    #     f.writelines(str(f) + "\n" for f in file_set)
+        # with open(f"{analyzed_root}_ownership_lineageos_{sha}.csv", "w", newline="") as f:
+        #     f.writelines(str(f) + "\n" for f in file_set)
 
-    # with open("blame_dict", "wb") as f:
-    #     pickle.dump(blame_dict, f)
-    end_time = time()
-    print(end_time - start_time)
+        # with open("blame_dict", "wb") as f:
+        #     pickle.dump(blame_dict, f)
+        end_time = time()
+        print(end_time - start_time)
 
 
 if __name__ == '__main__':
